huge overhaul

+++ app/Http/Controllers/API/BookmarkController.php
@@ -7,7 +7,7 @@ use App\Bookmark;
 use App\EDokumen;
 use App\Http\Controllers\Controller;
 use App\Konten;
-use App\Penulis;
+use App\User;
 use App\VideoAudio;
 use Illuminate\Support\Facades\Auth;
 
@@ -23,7 +23,7 @@ class BookmarkController extends Controller
                 'foto' => $temp->foto,
             ]);
 //        }
-        } elseif ($tipe == "VideoAudio") {
+        } elseif ($tipe == "Video") {
             $temp = VideoAudio::where('id', '=', $id)->first();
             $var = array([
                 'isi' => $temp->isi,
@@ -46,27 +46,35 @@ class BookmarkController extends Controller
 
     private function get_penulis($id)
     {
-        $var = Penulis::where('id', $id)->first();
-        $var->makeHidden(['id', 'created_at', 'updated_at']);
-        return $var;
+        $var = User::where('id', $id)->first();
+        $user = array([
+            'nama' => $var->nama,
+            'peran' => $var->peran,
+        ]);
+        return $user;
     }
 
     private function get_data($list)
     {
         $var = array();
         foreach ($list as $l) {
-            $konten = Konten::where('id', $l->id_konten);
-            $isi = $this->get_isi($konten->tipe, $konten->id_tipe);
-            $penulis = $this->get_penulis($konten->id_penulis);
-            $temp = array(
-                'id' => $konten->id,
-                'tipe' => $konten->tipe,
-                'judul' => $konten->judul,
-                'tanggal' => $konten->tanggal,
-                'konten' => $isi,
-                'penulis' => $penulis
-            );
-            array_push($var, $temp);
+            if ($l->is_hidden != 1){
+                $konten = Konten::where('id', $l->konten_id)->first();
+                $isi = $this->get_isi($konten->tipe, $konten->id_tipe);
+                $penulis = $this->get_penulis($konten->user_id);
+                $temp = array(
+                    'id' => $l->id,
+                    'konten_id' => $konten->id,
+                    'tipe' => $konten->tipe,
+                    'judul' => $konten->judul,
+                    'tanggal' => $konten->tanggal,
+                    'konten' => $isi,
+                    'penulis' => $penulis
+                );
+                array_push($var, $temp);
+            } else {
+                continue;
+            }
         }
         return $var;
     }
@@ -75,27 +83,29 @@ class BookmarkController extends Controller
     public function add($id)
     {
         try {
-            if ($user = Auth::guard('pakar_api')->user()) {
+            $user = Auth::guard('api')->user();
+
+            if (Bookmark::where('user_id', $user->id)->where('konten_id', $id)->exists()) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Anda sudah mem-bookmark konten ini',
+                    'Status' => 403
+                ], 403);
+            } else {
                 $var = Bookmark::create([
-                    'id_pakar_sawit' => $user->id,
-                    'id_konten' => $id,
+                    'user_id' => $user->id,
+                    'konten_id' => $id,
                 ]);
-            } elseif ($user = Auth::guard('petani_api')->user()) {
-                $var = Bookmark::create([
-                    'id_petani' => $user->id,
-                    'id_konten' => $id,
-                ]);
+
+                $konten = Konten::where('id', $id)->first();
+
+                return response()->json([
+                    'success' => true,
+                    'message' => 'Konten ter-Bookmark',
+                    'judul' => $konten->judul,
+                    'Status' => 200
+                ], 200);
             }
-
-            $konten = Konten::where('id', $id)->first();
-
-            return response()->json([
-                'success' => true,
-                'message' => 'Konten ter-Bookmark',
-                'judul' => $konten->judul,
-                'Status' => 200
-            ], 200);
-
         } catch (\Exception $e) {
             return response()->json([
                 'success' => false,
@@ -108,14 +118,25 @@ class BookmarkController extends Controller
     public function delete($id)
     {
         try {
-            $var = Bookmark::where('id', $id)->first();
-            $var->delete();
+            $user = Auth::guard('api')->user();
+//            dd($user->id);
+            if (Bookmark::where('user_id', $user->id)->where('konten_id', $id)->exists()){
+                $var = Bookmark::where('user_id', $user->id)->where('konten_id', $id)->first();
+                $var->delete();
 
-            return response()->json([
-                'success' => true,
-                'message' => 'Bookmark terhapus',
-                'Status' => 200
-            ], 200);
+                return response()->json([
+                    'success' => true,
+                    'message' => 'Bookmark terhapus',
+                    'Status' => 200
+                ], 200);
+            } else {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Bookmark tidak ada',
+                    'Status' => 403
+                ], 403);
+            }
+
 
         } catch (\Exception $e) {
             return response()->json([
@@ -126,14 +147,12 @@ class BookmarkController extends Controller
         }
     }
 
-    public function get_all(){
+    public function get_all()
+    {
         try {
-            if ($user = Auth::guard('pakar_api')->user()) {
-                $list = Bookmark::where('id_pakar_sawit', $user->id)->get();
-            } elseif ($user = Auth::guard('petani_api')->user()) {
-                $list = Bookmark::where('id_petani', $user->id)->get();
-            }
+            $user = Auth::guard('api')->user();
 
+            $list = $user->bookmark;
             $bookmark = $this->get_data($list);
 
             return response()->json([
